# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Public Toilets API",
    "version": "1.0.0",
    "description": "REST API specification for public toilets and their reviews",
    "contact": {
      "name": "Gabrijel Krizmanić",
      "email": "gabrijel.krizmanic@gmail.com"
    },
    "license": {
      "name": "CC0 1.0 Universal (Public Domain Dedication)",
      "url": "https://github.com/gabrijel0741/OR_public_toilets/blob/main/LICENSE"
    }
  },
  "paths": {
...
    "/api/v1/reviews": {
      "get": {
        "summary": "Fetch all existing reviews",
        "responses": {
          "200": {
            "description": "A list of all reviews",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string",
                      "example": "All reviews found"
                    },
                    "response": {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/Review"
                      }
                    }
                  },
                  "required": ["status", "message", "response"]
                }
              }
            }
          },
          "404": {
            "description": "No reviews have been found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Not found"
                    },
                    "message": {
                      "type": "string",
                      "example": "No reviews have been found"
                    },
                    "response": {
                      "type": "string",
                      "example": "null"
                    }
                  },
                  "required": ["status", "message", "response"]
                }
              }
            }
          },
          "500":{
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Service unavailable"
                    },
                    "message": {
                      "type": "string",
                      "example": "Interval server error"
                    },
                    "response": {
                      "type": "string",
                      "example": "null"
                    }
                  },
                  "required": ["status", "message", "response"]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new review",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/definitions/ReviewInput" } } }
        },
        "responses": {
          "201": {
            "description": "Successfully added a new review",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "OK"
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully added a new review"
                    },
                    "response": {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/Review"
                      }
                    }
                  },
                  "required": ["status", "message", "response"]
                }
              }
            }
          },
          "400": {
            "description": "Review could not be added",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Bad request"
                    },
                    "message": {
                      "type": "string",
                      "example": "Toilet with the given toilet_id doesn't exist"
                    },
                    "response": {
                      "type": "string",
                      "example": "null"
                    }
                  },
                  "required": ["status", "message", "response"]
                }
              }
            }
          },
          "500":{
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Service unavailable"
                    },
                    "message": {
                      "type": "string",
                      "example": "Interval server error"
                    },
                    "response": {
                      "type": "string",
                      "example": "null"
                    }
                  },
                  "required": ["status", "message", "response"]
                }
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
...
      "Review": {
        "type": "object",
        "properties": {
          "review_id": { "type": "integer" },
          "review_user": { "type": "string" },
          "review_user_rating": { "type": "integer" },
          "review_comment": { "type": "string" },
          "toilet_id": { "type": "integer" }
        }
      },
      "ReviewInput": {
        "type": "object",
        "required": ["review_user","review_user_rating","toilet_id"],
        "properties": {
          "review_user": { "type": "string" },
          "review_user_rating": { "type": "integer", "minimum": 1, "maximum": 5 },
          "review_comment": { "type": "string" },
          "toilet_id": { "type": "integer" }
        }
      },
...
    }
}